Add plug for session ID and mount for tracking
The session ID is fetched from the cookie if it exists or a new session
ID is generated, set in the cookie, and then stored in the session.

The tracking mount pulls the ID from the session and stores the live
view module name as an encrypted string to be sent to the client. The
module name is encrypted to prevent leaking internal information.